% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLegendEPA.R
\name{addLegendEPA}
\alias{addLegendEPA}
\title{Add legend to EPA leaflet map}
\usage{
addLegendEPA(map, colors, labels, sizes, opacity = 0.8)
}
\description{
Helper function in \code{sapphirine::EPAMap}. Adds legend for EPA monitoring stations to EPA leaflet map
}
