% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customEPAData.R
\name{customEPAData}
\alias{customEPAData}
\title{Function}
\usage{
customEPAData(
  startDate,
  endDate,
  shape = sapphirine::GPA_counties,
  variables = c("PM2.5", "PM10", "SO2", "NO2", "O3", "CO")
)
}
\description{
Function
}
\examples{
customEPAData()
}
