% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.top.R
\name{f.top}
\alias{f.top}
\title{Top value function}
\usage{
f.top(x)
}
\arguments{
\item{x}{The data values to be represented by the legend}
}
\description{
Helper function for \code{localMap} and \code{EPAMap}. Gives optimal top value for legends in leaflet map display
}
