% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.zoom.R
\name{f.zoom}
\alias{f.zoom}
\title{Zoom function}
\usage{
f.zoom(x, y)
}
\arguments{
\item{x}{The range of x coordinates (longitudes) of the display}

\item{y}{The range of y coordinates (latitudes)}
}
\description{
Helper function for \code{localMap} and \code{EPAMap}. Gives optimal zoom for initial display in leaflet maps
}
