% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intEPARaster.R
\name{intEPARaster}
\alias{intEPARaster}
\title{Function}
\usage{
intEPARaster(
  data,
  shape,
  power = 1,
  variables = c("PM2.5", "PM10", "SO2", "NO2", "O3", "CO")
)
}
\description{
Function
}
\examples{
intEPARaster()
}
