% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localMap.R
\name{localMap}
\alias{localMap}
\title{Local Raster Leaflet Map}
\usage{
localMap(
  rasBrick,
  popups = TRUE,
  bounds = NULL,
  colors = "YlOrRd",
  d.colors = "Purples"
)
}
\arguments{
\item{rasBrick}{An output of \code{sapphirine::localRaster}}

\item{popups}{Whether to include clickable popup displays with summary data for every raster bin in the map display.}

\item{bounds}{An optional shapefile input to project onto the map display.}

\item{colors}{Color palette for raster layers}

\item{d.colors}{Color palette for measurement density layers. Relevant only if \code{sapphirine::localRaster} output contains density plots.}
}
\description{
Render a \code{leaflet} map display for brick of local data rasters.
}
\examples{

#Select counties for shape
counties <- c('Bucks', 'Chester', 'Delaware', 'Montgomery', 'Philadelphia')
cty.shape <- selectGPACounties(counties)

#Customize subset of local.data
dat <- customLocalData(cty.shape, '2017-06-01', '2019-05-31')

#Make raster brick
localRaster(dat, cty.shape, 100, 100)

#Select Philadelphia shapefile to display Philadelphia boundaries in map
phl.bound <- selectGPACounties('Philadelphia')

localMap(ras, bounds = phl.bound)
}
