% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localRaster.R
\name{localRaster}
\alias{localRaster}
\title{Function}
\usage{
localRaster(
  data,
  shape,
  nrows,
  ncols,
  variables = c("Temperature", "Humidity", "PM1", "PM2.5", "PM10", "Crime", "Poverty",
    "Traffic"),
  includeCount = TRUE
)
}
\description{
Function
}
\examples{
localRaster()
}
