% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectGPACounties.R
\name{selectGPACounties}
\alias{selectGPACounties}
\title{Select Greater Philadelphia counties}
\usage{
selectGPACounties(counties)
}
\arguments{
\item{counties}{A vector containing names of Greater Philadelphia counties to be included. Proper names can be accessed from \code{sapphirine::GPACountyNames}}
}
\value{
A SpatialPolygonsDataFrame shapefile of the input counties.
}
\description{
Allows for easy selection of Greater Philadelphia Area counties to generate a custom shapefile.
}
\examples{
#Print a list of county names for reference
GPACountyNames

counties <- c('Bucks', 'Chester', 'Delaware', 'Montgomery', 'Philadelphia')

selectGPACounties(counties)
}
