% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic.raster.R
\docType{data}
\name{traffic.raster}
\alias{traffic.raster}
\title{Import Traffic Raster}
\format{
An object of class \code{RasterLayer} of dimension 1040 x 1825 x 1.
}
\usage{
traffic.raster
}
\description{
Import Traffic Raster
}
\keyword{datasets}
