% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customLocalData.R
\name{customLocalData}
\alias{customLocalData}
\title{Function}
\usage{
customLocalData(
  shape = sapphirine::GPA_counties,
  startDate,
  endDate,
  startTime = "00:00",
  endTime = "23:59",
  sensors = sapphirine::sensor.list
)
}
\description{
Function
}
\examples{
customLocalData()
}
