% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intEPARaster.R
\name{intEPARaster}
\alias{intEPARaster}
\title{Interpolated Raster from EPA Data}
\usage{
intEPARaster(
  data,
  shape,
  power = 1,
  variables = c("PM2.5", "PM10", "SO2", "NO2", "O3", "CO")
)
}
\arguments{
\item{data}{A subset of \code{sapphirine::EPA.data}. See \code{sapphirine::customLocalData} for easy custom subsetting.}

\item{shape}{A SpatialPolygons or extent object, for example, a subset of \code{sapphirine::GPA_counties} created using \code{sapphirine::selectGPACounties}, to define the boundaries of the raster layers.}

\item{power}{Power of inverse-distance-weighted interpolation (e.g., \code{power = 2} will call inverse-square-distance-weighted interpolation)}

\item{variables}{A vector of \code{sapphirine::EPA.data} variables to include.}
}
\description{
Create a brick of 1km x 1km raster layers rendered with inverse-distance-weighted interpolation of EPA data.
}
\examples{

#Select counties for shape
counties <- c('Bucks', 'Chester', 'Delaware', 'Montgomery', 'Philadelphia')
cty.shape <- selectGPACounties(counties)

#Customize subset of EPA.data
dat <- customEPAData(cty.shape, '2017-06-01', '2019-05-31')

intEPARaster(dat, cty.shape)
}
