% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localMap.R
\name{localMap}
\alias{localMap}
\title{Function}
\usage{
localMap(
  rasBrick,
  popups = TRUE,
  bounds = NULL,
  colors = "YlOrRd",
  d.colors = "Purples"
)
}
\description{
Function
}
\examples{
localMap()
}
