% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localRaster.R
\name{localRaster}
\alias{localRaster}
\title{Raster from Local Data}
\usage{
localRaster(
  data,
  shape,
  nrows,
  ncols,
  variables = c("Temperature", "Humidity", "PM1", "PM2.5", "PM10", "Crime", "Poverty",
    "Traffic"),
  includeCount = TRUE
)
}
\arguments{
\item{data}{A subset of \code{sapphirine::local.data}. See \code{sapphirine::customLocalData} for easy custom subsetting.}

\item{shape}{A SpatialPolygons or extent object, for example, a subset of \code{sapphirine::GPA_counties} created using \code{sapphirine::selectGPACounties}, to define the boundaries of the raster layers.}

\item{nrows}{Number of rows to include in the raster display}

\item{ncols}{Number of columns " "}

\item{variables}{A vector of \code{sapphirine::local.data} variables to include.}

\item{includeCount}{Whether to include density plot raster layers for sensor-based measurement variables (Temperature, Humidity, PM1, PM2.5, PM10) included in \code{variables}}
}
\value{
A RasterBrick object consisting of raster layers for all variables included.
}
\description{
Create a brick of raster layers corresponding to \code{sapphirine::local.data} variables.
}
\examples{

#Select counties for shape
counties <- c('Bucks', 'Chester', 'Delaware', 'Montgomery', 'Philadelphia')
cty.shape <- selectGPACounties(counties)

#Customize subset of local.data
dat <- customLocalData(cty.shape, '2017-06-01', '2019-05-31')

localRaster(dat, cty.shape, 100, 100)
}
