% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLabelFormat.R
\name{myLabelFormat}
\alias{myLabelFormat}
\title{Function}
\usage{
myLabelFormat(
  prefix = "",
  suffix = "",
  between = " &ndash; ",
  digits = 3,
  big.mark = ",",
  transform = identity,
  t.val = Inf
)
}
\description{
Function
}
\examples{
myLabelFormat()
}
